python/requirements.txt
v2xflexstack==0.11.0
numpy>=1.21
